Hide unprocessed within the cloner
